Rename file + replace options with enum values
